order_management/requirements.txt
celery==5.4.0
redis==5.0.1
Django==5.0.10
django-environ==0.11.2
djangorestframework==3.14.0
pydantic==1.8.2
typing_extensions==4.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
setuptools==58.0.4